refactor hook scrollToTop
